Tier-2 exactness vs double-precision oracle: 30-day roundoff-level shadowing
- Parameterize init-constant precision (PolarFilter factors/row extent,
  hsat/T1c tables): float64 default matches an r8 build and production;
  float32 mirrors the standard REAL*4 build. The filtered-row count sits
  on a truncation knife edge ((1-60/YB)*ny/2 == 5 exactly): REAL*4 gives
  js=4, REAL*8 js=5 - the two Fortran builds filter different rows.
- Golden tests now run against both oracles: f32 fixtures with f32-mode
  constants (documented f32 noise floors) and r8 fixtures at rtol 1e-11
  (measured floor ~1e-14). 99 tests.
- Fix AB3 history slice ordering in fixture consumers: Fortran bartr
  rotates slots write->oldest, so r_{n-1}=(w+1)%3, r_{n-2}=(w+2)%3
  (ascending guess is wrong one phase in three; found via a period-3
  bartr failure in a 72-step day bisection).
- Fix trajectory generator: step days manually with a single getbnd.
  m.qtcm() calls wgetbnd again, and a same-day second call on mid-month
  bracket days corrupts var_prior with next month's data (bndry1 guard
  restores time1/time2 but var_next has already advanced). This harness
  bug had injected a spurious 0.1 K kick into the control on Feb 15.
- tools: audit_vs_fixture.py (per-routine max-rel-diff table),
  gen_step_bisect.py (per-routine capture of every step of one day),
  fig_trajectory_r8.py; run_trajectory_port.py grows --mode f32|r8.
- tests/test_trajectory.py: 3-day race vs the r8 oracle (day-1 RMS
  ~1e-13 measured, thresholds 100x above).

Result: over 30 days (2160 steps) the port's RMS distance from the
bit-deterministic r8 Fortran stays at accumulated float64 roundoff
(T1 9.4e-13 K at day 30), within ~3x of an r8 twin seeded with a
1e-15 relative perturbation - dynamically indistinguishable from a
last-bit difference.

+++ tools/gen_golden.py
@@ -72,7 +72,7 @@ def make_capture(ext):
     def capture():
         state = {k: get_array(k) for k in ARRAY_KEYS}
         for k in SCALAR_REAL_KEYS:
-            state[k] = np.float32(sp.getitem_real(k))
+            state[k] = np.float64(sp.getitem_real(k))
         return state
 
     return capture
